Move all headers to the 'ferrix-widgets' crate
