+++ ferrix-app/src/pages/ram.rs
@@ -23,12 +23,10 @@
 use crate::{
     Message, fl,
     load_state::DataLoadingState,
-    widgets::{
-        header,
-        table::{InfoRow, fmt_val, kv_info_table},
-    },
+    widgets::table::{InfoRow, fmt_val, kv_info_table},
 };
 use ferrix_lib::ram::{RAM, Swaps};
+use ferrix_widgets::headers::header;
 use iced::widget::{column, container, scrollable, text};
 
 pub fn ram_page<'a>(
