+++ ferrix-widgets/src/lib.rs
@@ -21,5 +21,6 @@
 pub mod file_manager;
 // pub mod card;
 pub mod button;
+pub mod headers;
 pub mod icons;
 pub mod tooltip;
